+++ RELEASE-NOTES.md
@@ -4,6 +4,7 @@
 * #115 Table: loss of precision when inserting BigDecimal to Derby
 * #116 CsvDataSetBuilder can't handle BigDecimal
 * #121 Upgrade Derby to 10.15.2.0
+* #122 JUnit 5 connector - unify and expand test SQL conditions builders
 
 ## 2.0.B1